requirements.txt
Django==1.9.4
django-bootstrap3==7.0.1
djangorestframework==3.3.3
wheel==0.26.0
